[fix][chaejy]: 오타 수정
