+++ src/shared/ui/profile-image/ProfileImage.stories.tsx
@@ -5,13 +5,23 @@ const meta: Meta<typeof ProfileImage> = {
   title: 'Shared/UI/ProfileImage',
   component: ProfileImage,
   tags: ['autodocs'],
-};
-export default meta;
-
-type Story = StoryObj<typeof ProfileImage>;
-
-export const Default: Story = {
+  parameters: {
+    layout: 'centered',
+    controls: { include: ['size'] },
+  },
+  argTypes: {
+    size: {
+      options: ['s', 'm', 'l', 'xl'],
+      control: { type: 'radio' },
+    },
+  },
   args: {
+    alt: '프로필 이미지',
     size: 'l',
   },
 };
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {};
